refactor blueprint service for HA
Cleanup blueprints by moving the service into its own package.

Closes #297
