+++ platform/api/server/list/client.go
@@ -36,16 +36,6 @@ func NewClient(instance string, logger log.Logger, token string, opts ...httptra
 			opts...,
 		).Endpoint()
 	}
-	var getBlueprintsEndpoint endpoint.Endpoint
-	{
-		getBlueprintsEndpoint = httptransport.NewClient(
-			"GET",
-			copyURL(u, "/v1/blueprints"),
-			encodeRequestWithToken(token, EncodeHTTPGenericRequest),
-			DecodeGetBlueprintsResponse,
-			opts...,
-		).Endpoint()
-	}
 
 	var getDEPAccountInfoEndpoint endpoint.Endpoint
 	{
@@ -105,7 +95,6 @@ func NewClient(instance string, logger log.Logger, token string, opts ...httptra
 	return Endpoints{
 		ListDevicesEndpoint:       listDevicesEndpoint,
 		GetDEPTokensEndpoint:      getDEPTokensEndpoint,
-		GetBlueprintsEndpoint:     getBlueprintsEndpoint,
 		GetDEPAccountInfoEndpoint: getDEPAccountInfoEndpoint,
 		GetDEPDeviceEndpoint:      getDEPDeviceDetailsEndpoint,
 		GetDEPProfileEndpoint:     getDEPProfilesEndpoint,
